correct format of module
