+++ src/modules/Convolution/index.js
@@ -1,0 +1,4 @@
+module.exports = [
+    require('./Module'),
+    require('./info.json')
+]
